requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Deep Learning
torch>=2.0.0

# Explainability
shap>=0.42.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0
requests>=2.31.0
